Backend/Retrieval/requirements.txt
qdrant-client>=1.9.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0